Make enums class enums
Safer to use class enums and static cast to avoid potential bugs by implicit
casts.

+++ src/packet.cpp
@@ -21,7 +21,7 @@ Packet::Packet(const uint32_t payload_size) : size(payload_size + HEADER_SIZE) {
 
 Packet::Packet(const MessageType type, const std::string &msg)
     : Packet(msg.size()) {
-    this->header->type = type;
+    this->header->type = static_cast<uint8_t>(type);
     std::strncpy(this->payload, msg.c_str(), msg.size());
 }
